Start to light detection
A working version of a enemy light detection system has been made. This system will be used for enemies in the game. When an enemy is illuminated by light they will take damage. Currently the testing spheres display when they are or aren't in light based off their color.

This system works as follows. Once a gameobject enters the camera / is visible it gets added to a dictionary. For each visible enemy a bounding box of rays are casted from the camera. If a ray hits the enemy gameobject it then checks the pixels color value, (pixel from the render texture i.e lower res pixel size). If the brightness is higher than a threshold the enemy is in light. Performance is ok for the time being. 10 enemies can be on screen with 80+ fps. I will keep making performance improvements where possible.