Code Cleanup & Resolution Additions
Cleaned up the code of the player, moving the light detection code to a custom EnemyManager class. This class will be used to spawn / despawn enemies, track if they are in light or not based on visibility, etc.

Along side the code cleanup, a resolution manager class has been made. Making it so when the resolution changes the proper variables change. Making it so the resolution can change without breaking the game. Different aspect ratios from 16:9 will be added next.