Interact Interface & Start To Crafting
Interactable Interface implemented allowing player to interact with interactable objects. Interact (class for player) is responsible for checking if the player is hovering over a interactable object. Custom mouse over & mouse exit is implemented due to the render texture messing with unity's built in mouse over functionality.

Crafting fabricator start. Can be interacted with and displays when the cursor is hovering over it with a light. Once clicking it the camera view transitions to the fabricator view. This will be where the player moves items into the fabricator and crafts items.