Player Inventory - Items - EventManager
Player Inventory started. Currently each item used in the game will have a ScriptableObject containing all of said items data , such as name, model, max stack size, etc. Each item is stored into the PlayerInventory in a dictionary with the item being the key and the int being the value. The int represents how much of said item the player has.

A EventManager class has been created to allow for easy use of events between different scripts used in the project, allowing any script to listen or invoke an event as needed.