Crafting Fabricator Item Display
A display now comes into view once the player enters crafting mode. Here a board moves into view and displays what the player has in their inventory that can be used for crafting. The mesh of the item is placed in view to be interacted with, dragging / dropping for crafting. Assets very temp at the moment.

When an item is added or removed from the players inventory, the next time the player goes to craft the board will have the correct count of the affected item displayed.