Crafting Base
Base crafting functionality done, you can click and drag displayed crafting items around. By moving an item into the blue square (insert slot) you put an item in to craft. From here by pressing F3 it will craft (temporary, in the future you will pull the lever).

When a craft fails the items are returned back to the player inventory. And if a craft is correct it will place the crafted item into the players inventory.