PlayerEquipment + Item Improvement
PlayerEquipment will be used for items that the player can use to aid them. The first equipment item implemented is a flare. Using right click to use an equipment item, in this case the flare, will use said equipment item. With the flare it is thrown in the direction the player is looking, after a set amount of time the light goes out. Then later it is destroyed.

Item improvement for displaying items in the world. Now instead of using four point lights to equally light an item in the world for the player to pick up, now only one light is used. Each time the player rotates the camera, and event is called for each item to rotate their light as well. This way the item is always visible from the players perspective.